User Authentication

+++ frontend/src/components/PlaylistDetails.js
@@ -1,4 +1,5 @@
 import { usePlaylistsContext } from '../hooks/usePlaylistsContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 // date fns
 import format from 'date-fns/format'
@@ -6,10 +7,18 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const PlaylistDetails = ({ playlist }) => {
     const { dispatch } = usePlaylistsContext()
+    const { user } = useAuthContext()
 
     const handleClick = async () => {
+        if (!user) {
+            return
+        }
+
         const response = await fetch('/api/playlists/' + playlist._id, {
-            method: "DELETE"
+            method: "DELETE",
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
